* 3.0-videos:
  Fix .slug initializer
  Use scriptBottom block
  Fix indentation
  Add support for external video links
  Enable filtering album photos by by media type
  Remove hardcoded icon codes and use map
  Create appropriate thumbnail based on mime_type
  Include ffmpeg-php lib in Vendor/ffmpeg-php
  Add support for some video formats
  Fix album index link
  Remove span12 and use width: 100% instead in #return style
  Minor formatting and docblock boilerplate

+++ View/Albums/admin_upload.ctp
@@ -20,7 +20,7 @@ $editUrl = $this->Html->link($album['Album']['title'], array(
 ));
 
 $this->Html
-	->addCrumb('', '/admin', array('icon' => 'home'))
+	->addCrumb('', '/admin', array('icon' => $_icons['home']))
 	->addCrumb('Gallery')
 	->addCrumb(__d('gallery', 'Albums'), array(
 		'admin' => true,
@@ -40,6 +40,19 @@ else:
 endif;
 
 $this->start('actions');
+	echo $this->Html->link(__d('gallery', 'Add Video Link'),
+		array(
+			'plugin' => 'gallery',
+			'controller' => 'photos',
+			'action' => 'add',
+			'?' => array(
+				'album_id' => $album['Album']['id'],
+			),
+		),
+		array(
+			'button' => 'default',
+		)
+	);
 	echo $this->Form->postLink(__d('gallery', 'Reset weight'),
 		array(
 			'plugin' => 'gallery',
@@ -66,33 +79,46 @@ endif;
 $this->append('main');
 ?>
 	<div id="upload" class="<?php echo $columnFull; ?>"></div>
-	<div id="return" class="<?php echo $columnFull; ?>">
+	<div id="return">
 
 		<?php foreach($album['Photo'] as $photo): ?>
 			<div class="album-photo">
 				<div class="preview">
 				<?php
-				echo $this->Html->link(
-					$this->Html->thumbnail($photo['ThumbnailAsset']['path']),
-					$photo['LargeAsset']['path'],
-					array(
-						'rel' => 'gallery-' . $photo['AlbumsPhoto']['album_id'],
-						'class' => 'thickbox',
-						'escape' => false,
-					)
-				);
+				if ($photo['media_type'] == MediaType::VIDEO_LINK):
+					echo $this->Gallery->previewExternalVideo($photo['external_url']);
+				else:
+					$displayPath = isset($photo['LargeAsset']['path']) ? $photo['LargeAsset']['path'] : $photo['OriginalAsset']['path'];
+					echo $this->Html->link(
+						$this->Html->thumbnail($photo['ThumbnailAsset']['path']),
+						$displayPath,
+						array(
+							'rel' => 'gallery-' . $photo['AlbumsPhoto']['album_id'],
+							'class' => 'thickbox',
+							'escape' => false,
+						)
+					);
+				endif;
 				?>
 				</div>
 
 				<div class="path">
 				<?php
-					$filename = basename($photo['LargeAsset']['path']);
+				if ($photo['media_type'] == MediaType::VIDEO_LINK):
+					$url = $photo['external_url'];
+					$link = $this->Html->link($url, $url, array(
+						'target' => '_blank',
+					));
+					echo __d('gallery', 'URL: %s', $link);
+				else:
+					$filename = basename($displayPath);
 					$filename = $this->Html->link(
 						$this->Text->truncate($filename, 120),
 						$photo['OriginalAsset']['path'],
 						array('target' => '_blank', 'title' => $filename)
 					);
 					echo __d('gallery', 'Filename: %s', $filename);
+				endif;
 				?>
 				</div>
